+++ packages/features/ee/teams/components/AddNewTeamMembers.tsx
@@ -50,6 +50,7 @@ const AddNewTeamMembers = ({ isOrg = false }: { isOrg?: boolean }) => {
   if (session.status === "loading" || !teamQuery.data) return <AddNewTeamMemberSkeleton />;
 
   return (
+    // TODO: Get members
     <AddNewTeamMembersForm
       defaultValues={{ members: teamQuery.data.members }}
       teamId={teamId}
