+++ packages/features/ee/teams/components/MakeTeamPrivateSwitch.tsx
@@ -25,7 +25,6 @@ const MakeTeamPrivateSwitch = ({
     },
     async onSuccess() {
       await utils.viewer.teams.get.invalidate();
-      await utils.viewer.teams.getMinimal.invalidate();
       showToast(t(isOrg ? "your_org_updated_successfully" : "your_team_updated_successfully"), "success");
     },
   });
