+++ packages/features/ee/teams/components/TeamEventTypeForm.tsx
@@ -27,10 +27,7 @@ export const TeamEventTypeForm = ({
 }: props) => {
   const { t } = useLocale();
   const orgBranding = useOrgBranding();
-  const { data: team } = trpc.viewer.teams.getMinimal.useQuery(
-    { teamId, isOrg: false },
-    { enabled: !!teamId }
-  );
+  const { data: team } = trpc.viewer.teams.get.useQuery({ teamId, isOrg: false }, { enabled: !!teamId });
   const urlPrefix = orgBranding?.fullDomain ?? process.env.NEXT_PUBLIC_WEBSITE_URL;
 
   const { form, createMutation, isManagedEventType } = useCreateEventType(onSuccessMutation, onErrorMutation);
