+++ packages/features/ee/teams/pages/team-profile-view.tsx
@@ -103,7 +103,7 @@ const ProfileView = ({ isAppDir }: { isAppDir?: boolean }) => {
     data: team,
     isPending,
     error,
-  } = trpc.viewer.teams.getMinimal.useQuery(
+  } = trpc.viewer.teams.get.useQuery(
     { teamId },
     {
       enabled: !!teamId,
@@ -262,7 +262,7 @@ const ProfileView = ({ isAppDir }: { isAppDir?: boolean }) => {
   );
 };
 
-export type TeamProfileFormProps = { team: RouterOutputs["viewer"]["teams"]["getMinimal"] };
+export type TeamProfileFormProps = { team: RouterOutputs["viewer"]["teams"]["get"] };
 
 const TeamProfileForm = ({ team }: TeamProfileFormProps) => {
   const utils = trpc.useUtils();
