+++ packages/trpc/server/routers/viewer/teams/get.schema.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
-export const ZGetInputSchema = z.object({
+export const ZGetSchema = z.object({
   teamId: z.number(),
   isOrg: z.boolean().optional(),
 });
 
-export type TGetInputSchema = z.infer<typeof ZGetInputSchema>;
+export type TGetInputSchema = z.infer<typeof ZGetSchema>;
