--- packages/trpc/server/routers/viewer/teams/getMinimal.schema.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const ZGetMinimalSchema = z.object({
-  teamId: z.number(),
-  isOrg: z.boolean().optional(),
-});
-
-export type TGetMinimalInputSchema = z.infer<typeof ZGetMinimalSchema>;
